// src/app/Interfaces/DashboardAssetStatus.ts
export class DashboardAssetStatus {
    AssetType: string;
    TotalAssetCount: number;
    InUseAssetCount: number;
    IdleAssetCount: number;
    AlertAssetCount: number;
    OnlineAssetCount: number;
    OfflineAssetCount: number;
    IndoorAssetCount: number;
    OutDoorAssetCount: number;
    IndoorOutDoorAssetCount: number;
      TotalFacilityCount: number;
      CurrentIndoorAlertCount: number;
      CurrentOutdoorAlertCount: number;

}